Use backing fields & public property instead of function to expose LiveData
There is no functional difference - either way MutableLiveData is not
exposed outside of the ViewModel class. Using backing field & property
seems to be more commonly used than the private property MutableLiveData
and function LiveData initially used (which is used in ViewModel
documentation:
https://developer.android.com/topic/libraries/architecture/viewmodel)

In interview, it was suggested to use single property - still unclear
why this would be preferable since it would require casting to
MutableLiveData. Either I didn't understand what the suggestion was, or it
was a test to see if I would see the disadvantages of that approach?