Move logic of # of repos to display to ViewModel
Previously, RepoListAdapter (the RecyclerView.Adapter) observed changes
to all repos owned by an org. Also, the most starred repos for an org
was frequently recalculated from the complete list.

Now, the logic for which repos to displayed has been moved into the
ViewModel where it belongs, and RepoListAdapter observes the topRepos,
not allRepos.